rf: make integration test directory
