+++ tests/mod.rs
@@ -1,0 +1,2 @@
+mod status;
+mod submit;
