+++ tests/submit/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "haskell")]
+mod haskell;
